refactor: Update the useUniqueId hook
