+++ src/components/FormRenderer/components/Checkbox/index.tsx
@@ -19,8 +19,8 @@ import { Option } from '../../types';
 import Checkbox from '../../../Checkbox';
 import FormField from '../../../FormField';
 import Stack from '../../../../layouts/Stack';
-import { getControlId } from '../../getContolId';
-import { getErrorText } from '../../getErrorText';
+import useUniqueId from '../../../../hooks/useUniqueId';
+import { getErrorText } from '../../utils/getErrorText';
 
 interface CheckboxMappingProps {
     option: Option;
@@ -49,7 +49,7 @@ const CheckboxGroupMapping: FunctionComponent<UseFieldApiConfig> = (props) => {
         showError,
         meta: { error, submitFailed },
     } = useFieldApi(props);
-    const controlId = getControlId(input.name);
+    const controlId = useUniqueId(input.name);
     const errorText = getErrorText(validateOnMount, submitFailed, showError, error);
     if (options?.length > 0) {
         return (
