+++ src/components/FormRenderer/components/Datepicker/index.tsx
@@ -17,8 +17,8 @@ import React, { FunctionComponent } from 'react';
 import useFieldApi, { UseFieldApiConfig } from '@data-driven-forms/react-form-renderer/use-field-api';
 import FormField from '../../../FormField';
 import Datepicker from '../../../DatePicker';
-import { getControlId } from '../../getContolId';
-import { getErrorText } from '../../getErrorText';
+import useUniqueId from '../../../../hooks/useUniqueId';
+import { getErrorText } from '../../utils/getErrorText';
 
 const DatepickerMapping: FunctionComponent<UseFieldApiConfig> = (props) => {
     const {
@@ -35,7 +35,7 @@ const DatepickerMapping: FunctionComponent<UseFieldApiConfig> = (props) => {
         showError,
         meta: { error, submitFailed },
     } = useFieldApi(props);
-    const controlId = getControlId(input.name);
+    const controlId = useUniqueId(input.name);
     const errorText = getErrorText(validateOnMount, submitFailed, showError, error);
     return (
         <FormField
