--- src/components/FormRenderer/getContolId/index.ts
@@ -1,22 +1,0 @@
-/** *******************************************************************************************************************
-  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-  
-  Licensed under the Apache License, Version 2.0 (the "License").
-  You may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  
-      http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.                                                                              *
- ******************************************************************************************************************** */
-import { v4 as uuidv4 } from 'uuid';
-
-const getControlId = (name: string) => {
-    return name ?? uuidv4();
-};
-
-export { getControlId };
